Conflicts:
	Gemfile
	Gemfile.lock
	Gemfile_rails4.lock
	README.md
	app/assets/javascripts/discourse/routes/application_route.js
	app/controllers/users/omniauth_callbacks_controller.rb
	app/views/layouts/application.html.erb
	config/environments/production.rb
	config/initializers/sidekiq.rb
	lib/auth.rb

+++ app/models/category_featured_topic.rb
@@ -5,18 +5,25 @@ class CategoryFeaturedTopic < ActiveRecord::Base
   # Populates the category featured topics
   def self.feature_topics
     transaction do
-      Category.all.each do |c|
-        feature_topics_for(c)
-        CategoryFeaturedUser.feature_users_in(c)
+      current = {}
+      CategoryFeaturedTopic.select(:topic_id, :category_id).order(:rank).each do |f|
+        (current[f.category_id] ||= []) << f.topic_id
+      end
+      Category.select(:id, :topic_id).all.each do |c|
+
+        feature_topics_for(c, current[c.id] || [])
+        CategoryFeaturedUser.feature_users_in(c.id)
       end
     end
   end
 
-  def self.feature_topics_for(c)
+  def self.feature_topics_for(c, existing=nil)
     return if c.blank?
 
-    query = TopicQuery.new(self.fake_admin, per_page: SiteSetting.category_featured_topics, except_topic_ids: [c.topic_id], visible: true)
-    results = query.list_category(c).topic_ids.to_a
+    query = TopicQuery.new(self.fake_admin, per_page: SiteSetting.category_featured_topics, except_topic_ids: [c.topic_id], visible: true, no_definitions: true)
+    results = query.list_category(c).topic_ids
+
+    return if results == existing
 
     CategoryFeaturedTopic.transaction do
       CategoryFeaturedTopic.delete_all(category_id: c.id)
